Assignment with SHA 256 authentication completed

+++ src/main/java/resource/InventoryResource.java
@@ -22,13 +22,6 @@ public class InventoryResource {
                 return new InventoryResource();
         }
 
-        @Path("/hello")
-        @GET
-        @Produces(MediaType.TEXT_PLAIN)
-        public String sayPlainTextHello() {
-            return "Hello Inventory";
-        }
-
         // Object to JSON
         @Path("/{inventoryId}")
         @GET
@@ -49,32 +42,24 @@ public class InventoryResource {
         @Path("/list")
         @GET
         @Produces(MediaType.APPLICATION_JSON)
-        public Response getAllInventory(@QueryParam("category") String category,@QueryParam("location") String location,@HeaderParam("authorization") String header)
+        public Response getAllInventory(@QueryParam("category") String category,@QueryParam("location") String location)
         {
                 try
                 {
-                        if(inventoryService.validateUser(header))
-                        {
-                             if(category==null && location==null)
-                                     return Response.status(Response.Status.OK).entity(inventoryService.getInventoryAll()).build();
-                             else if(category!=null && location==null)
-                                     return Response.status(Response.Status.OK).entity(inventoryService.getInventoryByCategory(category)).build();
-                             else if(category==null && location!=null)
-                                     return Response.status(Response.Status.OK).entity(inventoryService.getInventoryByLocation(location)).build();
-                             else
-                                     return Response.status(Response.Status.OK).entity(inventoryService.getInventoryByCategoryAndLocation(category,location)).build();
-                        }
+                        if(category==null && location==null)
+                                return Response.status(Response.Status.OK).entity(inventoryService.getInventoryAll()).build();
+                        else if(category!=null && location==null)
+                                return Response.status(Response.Status.OK).entity(inventoryService.getInventoryByCategory(category)).build();
+                        else if(category==null && location!=null)
+                                return Response.status(Response.Status.OK).entity(inventoryService.getInventoryByLocation(location)).build();
                         else
-                        {
-                                String message = "{\"error_message\" : \"user not authorized\"}";
-                                return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
-                        }
+                                return Response.status(Response.Status.OK).entity(inventoryService.getInventoryByCategoryAndLocation(category,location)).build();
 
                 }
                 catch(Exception exc)
                 {
                         String message = "{\"error_message\" : \"invalid id in request\"}";
-                        return Response.status(Response.Status.BAD_REQUEST).entity(exc.getMessage()).build();
+                        return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
                 }
         }
 
@@ -86,20 +71,18 @@ public class InventoryResource {
         public Response addNewInventoryItem(Inventory inventory,@HeaderParam("authorization") String header) {
                 try
                 {
-                        System.out.println("Header:" + header);
                         if(inventoryService.validateUser(header))
                                 return Response.status(Response.Status.OK).entity(inventoryService.addItem(inventory)).build();
                         else
                         {
-                                String message = "{\"error_message\" : \"user not authorized\"}";
+                                String message = "{\"message\" : \"Unauthorized user\"}";
                                 return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
                         }
                 }
                 catch(Exception exc)
                 {
                         String message = "{\"error_message\" : \"invalid payload\"}";
-                        System.out.println(exc.getMessage());
-                        return Response.status(Response.Status.BAD_REQUEST).entity(exc.getMessage()).build();
+                        return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
                 }
         }
 
@@ -107,11 +90,19 @@ public class InventoryResource {
         @Path("/update/{inventoryId}")
         @Produces(MediaType.APPLICATION_JSON)
         @Consumes(MediaType.APPLICATION_JSON)
-        public Response updateItem(Inventory inventory,@PathParam("inventoryId") String inventoryId)
+        public Response updateItem(Inventory inventory,@PathParam("inventoryId") String inventoryId,@HeaderParam("authorization")String header)
         {
                 try
                 {
-                        return Response.status(Response.Status.OK).entity(inventoryService.updateInventory(inventory,inventoryId)).build();
+                        if(inventoryService.validateUser(header))
+                        {
+                                return Response.status(Response.Status.OK).entity(inventoryService.updateInventory(inventory,inventoryId)).build();
+                        }
+                        else
+                        {
+                                String message = "{\"message\" : \"Unauthorized user\"}";
+                                return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
+                        }
                 }
                 catch(Exception exc)
                 {
@@ -123,17 +114,26 @@ public class InventoryResource {
         @DELETE
         @Path("/{inventoryId}/")
         @Produces(MediaType.APPLICATION_JSON)
-        public Response deleteItem(@PathParam("inventoryId") String inventoryId)
+        public Response deleteItem(@PathParam("inventoryId") String inventoryId,@HeaderParam("authorization") String header)
         {
                 try{
-                        inventoryService.deleteInventory(inventoryId);
-                        String message = "{\"message\" : \"OK\"}";
-                        return Response.status(Response.Status.OK).entity(message).build();
+                        if(inventoryService.validateUser(header))
+                        {
+                                String message = "{\"message\" : \"OK\"}";
+                                inventoryService.deleteInventory(inventoryId);
+                                return Response.status(Response.Status.OK).entity(message).build();
+                        }
+                        else
+                        {
+                                String message = "{\"error_message\" : \"Unauthorized user\"}";
+                                return Response.status(Response.Status.BAD_REQUEST).entity(message).build();
+                        }
+
                 }
                 catch(Exception exc)
                 {
                         String message = "{\"error_message\" : \"invalid id in request\"}";
-                        return Response.status(Response.Status.OK).entity(exc.getMessage()).build();
+                        return Response.status(Response.Status.OK).entity(message).build();
                 }
         }
 }
